qianwen_version/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
weaviate-client>=3.25.3
dashscope>=1.14.1
python-multipart>=0.0.9
PyMuPDF>=1.23.26
requests>=2.31.0
numpy>=1.26.0
tqdm>=4.66.1
pytesseract>=0.3.10
pdf2image>=1.16.3
python-docx>=0.8.11
Pillow>=10.2.0